Generating Triplets just fine, with caveats
But made some silly assumptions, like that a < b < c means something to an
algorithm. It really doesn't since a and b can just be swapped. This was a
naive assumption based on reading the problem too literally.

Now it's time to clean it up and get it to generate the correct answer.